+++ lib/queen_validators.dart
@@ -3,8 +3,7 @@ library queen_validators;
 /// `qValidattor` function
 export 'package:queen_validators/src/validator.dart';
 
-/// `abstract` `class` to impelemnt your own DTO || rule
-export 'package:queen_validators/src/imp/dto.dart';
+/// `abstract` `class` to impelemnt your own  rule
 export 'package:queen_validators/src/imp/validation.dart';
 
 /// export aviliable rules
@@ -12,6 +11,3 @@ export 'package:queen_validators/src/rules/index.dart';
 
 /// export aviliable extensions methods to validate variables
 export 'package:queen_validators/src/extensions/index.dart';
-
-/// only supported exception which will be trown in case of invalid DTO
-export 'package:queen_validators/src/exception.dart';