+++ lib/src/exception.dart
@@ -1,7 +1,0 @@
-/// impelementopn of `Exception` class
-/// but with another extra variable `erros`
-/// which will conatain the faild validation message from the `DTO`
-class QValidationException implements Exception {
-  final Map<String, List<String>> errors;
-  QValidationException(this.errors);
-}