+++ lib/src/rules/contains_any.dart
@@ -12,9 +12,6 @@ class ContainsAny extends QueenValidationRule {
 
   @override
   bool validate(String val) {
-    return this
-        .value
-        .where((element) => val.trim().contains(element))
-        .isNotEmpty;
+    return value.where((element) => val.trim().contains(element)).isNotEmpty;
   }
 }