+++ lib/src/rules/index.dart
@@ -5,7 +5,6 @@ export 'package:queen_validators/src/rules/is_in.dart';
 export 'package:queen_validators/src/rules/is_not_in.dart';
 export 'package:queen_validators/src/rules/is_optional.dart';
 export 'package:queen_validators/src/rules/is_required.dart';
-export 'package:queen_validators/src/rules/is_url.dart';
 export 'package:queen_validators/src/rules/match.dart';
 export 'package:queen_validators/src/rules/max_length.dart';
 export 'package:queen_validators/src/rules/min_length.dart';